Move updateView() from ViewsRepo to ViewDataRepo
